Update README.md
1. Correct the details of Kiwmi under ## Compositors
I submitted Kiwmi with details "A wayland compositor fully programmable with Lua" but it can also be configured like 'dwm' by
changing the header files in source. So corrected it to just "A fully programmable Wayland compositor", that's how its creator
describes it anyway.

2. Addded new compositor SWC
It has a small and easy to follow codebase. So it might be helpful for people wanting to get into Wayland compositor development.