Add Librewolf as a firefox based web browser.
Librewolf is a fork of Firefox, it is for users who don't want to sacrifice Internet speed using tor, but still wants privacy and freedom. Below is the introduction taken from https://librewolf.net/

This project is an independent fork of Firefox, with the primary goals of privacy, security and user freedom.

LibreWolf is designed to increase protection against tracking and fingerprinting techniques, while also including a few security improvements. This is achieved through our privacy and security oriented settings and patches. LibreWolf also aims to remove all the telemetry, data collection and annoyances, as well as disabling anti-freedom features like DRM.